backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
django-filter==23.5
django-ratelimit==4.1.0
django-cryptography==1.0
django-otp==2.0.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
django-logging==0.9.3
sentry-sdk==1.40.0
django-prometheus==2.3.1
tensorflow==2.15.0
pytest==7.4.3
pytest-django==4.5.2
djangorestframework-simplejwt==5.3.0
ccxt==4.2.0
pyotp
boto3
pytest
pytest-django
django-prometheus
sentry-sdk
drf-spectacular
redis
celery
